More json fixes
