+++ src/Microsoft.CodeAnalysis.CompilerDeveloperSdk/Protocol/Converters/DocumentUriConverter.cs
@@ -1,60 +1,21 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Roslyn.LanguageServer.Protocol
+using System;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+
+namespace Roslyn.LanguageServer.Protocol;
+
+/// <summary>
+/// TODO: document.
+/// </summary>
+internal class DocumentUriConverter : JsonConverter<Uri>
 {
-    using System;
-    using System.Globalization;
-    using Microsoft.CodeAnalysis.LanguageServer;
-    using Microsoft.CommonLanguageServerProtocol.Framework;
-    using Newtonsoft.Json;
-    using Newtonsoft.Json.Linq;
+    public override Uri Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+    => new(reader.GetString()!);
 
-    /// <summary>
-    /// TODO: document.
-    /// </summary>
-    internal class DocumentUriConverter : JsonConverter
-    {
-        /// <inheritdoc/>
-        public override bool CanConvert(Type objectType)
-        {
-            return true;
-        }
-
-        /// <inheritdoc/>
-        public override object? ReadJson(JsonReader reader, Type objectType, object? existingValue, JsonSerializer serializer)
-        {
-            reader = reader ?? throw new ArgumentNullException(nameof(reader));
-            if (reader.TokenType == JsonToken.String)
-            {
-                var token = JToken.ReadFrom(reader);
-                var uri = new Uri(token.ToObject<string>()!);
-
-                return uri;
-            }
-            else if (reader.TokenType == JsonToken.Null)
-            {
-                return null;
-            }
-
-            throw new JsonSerializationException(string.Format(CultureInfo.InvariantCulture, "Unable to deserialize Uri. Unexpected value encountered: {0}", reader.Value));
-        }
-
-        /// <inheritdoc/>
-        public override void WriteJson(JsonWriter writer, object? value, JsonSerializer serializer)
-        {
-            writer = writer ?? throw new ArgumentNullException(nameof(writer));
-
-            if (value is Uri uri)
-            {
-                var token = JToken.FromObject(uri.AbsoluteUri);
-                token.WriteTo(writer);
-            }
-            else
-            {
-                throw new ArgumentException($"{nameof(value)} must be of type {nameof(Uri)}");
-            }
-        }
-    }
+    public override void Write(Utf8JsonWriter writer, Uri value, JsonSerializerOptions options)
+        => writer.WriteStringValue(value.AbsoluteUri);
 }
